Add null handling for Books View

+++ BookishWebApplication/Models/Database/Book.cs
@@ -6,7 +6,7 @@ namespace BookishWebApplication.Models.Database
     {
         public int Id { get; set; }
         public string Title { get; set; }
-        public int PublicationYear { get; set; }
+        public short PublicationYear { get; set; }
         public string Isbn { get; set; }
     }
 }
